requirements.txt
# Databricks and Unity Catalog dependencies
databricks-sql-connector>=2.9.0
databricks-cli>=0.218.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
python-dotenv>=0.19.0
pydantic>=1.9.0

# Logging and development
structlog>=22.1.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0